backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.1.0
requests==2.28.1
feedparser==6.0.10
transformers==4.24.0
torch==1.13.1
sentencepiece==0.1.97
Pillow==9.3.0
pytesseract==0.3.10
opencv-python-headless==4.6.0.66
mediapipe==0.9.0.1
numpy==1.23.5